==> src/00_setup.R <==


install.packages(c("raster", "sf", "tidyverse", "RStoolbox",
                   "caret", "randomForest", "ggpubr", "mgcv", "Metrics",
                   "tidymodels", "rpart", "rpart.plot", "parttree"))
